TASK2 Consistency done,
Master is able to tell the client if a worker
has been added or deleted.

Every time client does a task checks if the master has changed and
updates its proxy list if it's necessary.

The workers now need the file name to perform any action, if the file
name has changed or it's null the worker updates its pandas datatframe.